rework to support multiple sub-commands
This reworks the CLI to take many sub-commands where the command is
specified as the first arg to the main `ggem` cli.  This also ports
the existing behavior as the first sub-command: `generate` (aliased
as `g`.

Note: this also updates some style things to get the code and tests
up to our modern conventions.