cli: add support for command exit errors
This is a custom exception that tells the CLI to just exit with
a non-zero status but don't output any error or help info.  This
isn't needed by any specific sub-commands but should be something
available to all sub commands.

This was needed by an Ardb sub command so we are making it a standard
for our sub command pattern/framework.