model a gemspec object
This models a gemspec.  It looks for a gemspec in a given root dir,
loads it and then can perform system commands on it.  This will be
used by the CLI to implement build, install and push sub-commands.

The `run_{build|install|push}_cmd` methods were all inspired by
Bundler's gem tasks.  However, instead of ripping off bundler's
implementation, I chose to have the commands just run the raw
system commands you could run manually.  This essentially means that
ggem's CLI commands will just be macros/aliases for the raw system
commands.  The end result, however, is the same as if you ran
bundler's corresponding rake tasks.

The goal with all of this is to provide an alternative to bundler's
rake tasks that doesn't rely on crappy Rake.  We are trying to
rid ourselves of all Rake dependency and this is part of that effort.

Note: this brings in the latest Scmd which has utils for spying and
testing Scmd calls.