model a git repo; add init cmd behavior; use in template generate
This models a git repo and the associated system cmds to initialize
it.  The init behavior is then applied to the generate command
logic and the previous template init logic is removed as it is
no longer needed.

The git repo will also be used in some future CLI sub commands.
Specifically tagging and push-with-tags logic will be used in a
coming gem "release" sub command.

Note: this also cleans up a few things on the gemspec model that
become apparent when adding the git repo model:

* remove the tap block arg - it isn't used and was incorrect
* remove the command specific exceptions and usage - there really
  isn't any value to them right now and they weren't being used and
  just added technical debt