add 'tag' CLI sub command
The command tags the commit with the spec's version information and
pushed any commits/tags.  This will be part of the coming 'release'
macro sub command that will tag the gem version and push the gem
to the push host.

This command requires the full string to tag the commit with.  I
chose to call the the `version_tag` and added this info to the
gemspec model so it would be centrally located alongside the raw
version info.

The tag command is both a git repo command and a gemspec command
in that it operates on a git repo object for the gem using the
spec's information.  I had to rework a couple things to make this
work as this is the first sub command that is both a repo command
and a spec command:

* I renamed the "execute" stuff to "notify cmd" naming.  This better
  describes the functions as they are about running system commands
  and notifying the CLI stdout/stderr.
* I broke out the base command running logic into a `cmd` function
  b/c the tag command has to do manual stdout/stderr notification
  b/c it doing a bunch of small system commands and needs very custom
  orchestration of those system commands and the related output to
  the CLI.
* I had to break out some CLI test contexts into mixins b/c the
  tag command had to mixin both the git repo spy and the gemspec spy
  contexts as it requires both.  I updated the existing tests
  accordingly.