git repo: add API for running tag and push related system commands
These commands will be used in the comming "tag" CLI sub command.
These methods just call the raw system commands to validate, tag
and push on the repo.

Note: I reworked the cmd tests in the gemspec tests to try an commonize
a bunch of boilerplate cmd test logic that I also needed for the
git repo tests.