add 'release' CLI sub command
This command is a macro command for running the tag then the push
sub commands (ie `ggem tag && ggem push`).  It follows that its
implementation would simply be to build tag/push command objects
and then run them.

This is the final sub command needed to replace bundler's gem helpers
rake commands.  The motivation for adding all of these is to be free
of needing (crappy) Rake as a dependency to use these nice gem handling
functions.