make the gem system tests more robust
This updates the system tests to make them more like our typical
tests and removes the use of assert macros in favor of custom
assertion methods.  This allows not doing all logic at the singleton
context scope.

This also updates the system tests to test that the gem `save!` method
returns self.  I accidentally removed this in a previous effort but
no tests failed to let me know.  Since all save testing happens in
the system tests, reworking the tests allowed me to easily add this
new assertion to the save tests.