hotfix: tweak some CLI help messages and the push command output
The help message tweaks are an attempt to make them more readable
and understandable.  The push command output change is to provide
immediate feedback that the push has been initiated while waiting
on the push to process.

These are just some minor UX tweaks from testing it out on a live
gem.

/cc @jcredding