cli: rework the install/push commands to compose the build command
This makes their implementation similar to the release command
which composes both the tag and the push commands.  It also removes
the need for the `notify_build` helper method.  This implementation
will scale better b/c any behavior or output added to the build
command will be used in the install/push commands.  This implementation
is the intent.

This also updates some tests for some CLI output and help messages.
These should have been done in the previous hotfix but were missed
b/c I forgot to run the test suite.