cli: reorg command definitions and CLIRB
This is a trivial reorganization of the CLI implementation code.
The goal here is to better organize things to not have to do the
"empty class" hack.  This also better abstracts the "implementation
details" of the CLI and highlights the business logic better.

This is also to prep for switching to using a custom class to hold
the set of commands available on the CLI (instead of a simple hash).
The reason for that coming effort is to have more control how the
available commands are displayed in the invalid commands help output.
I want to control the order and also better show how the "alias"
commands are related to the "real" commands.  I need a class to
insert some custom logic and provide a custom API for adding new
commands to accomplish these goals.  Anyway, that will be coming
in a future effort.