cli: new command config api; rich cmd lists for invalid help msgs
This reworks the API used to configure commands on the CLI.  The
overall goal here is to provide rich cmd lists for the general CLI
`--help` flag.

Here is a before/after of the `--help` output:

```
(before)
$ ggem -h
Usage: ggem [COMMAND] [options]

Commands: build, g, generate, install, push
Options:
        --version
        --help
```

```
(after)
$ ggem -h
Usage: ggem [COMMAND] [options]

Options:
        --version
        --help

Commands:
  generate (g) # Create a gem given a GEM-NAME
  build    (b) # Build ggem-1.7.0.gem into the pkg directory
  install  (i) # Build and install ggem-1.7.0.gem into system gems
  push     (p) # Push built ggem-1.7.0.gem to https://rubygems.org
  tag      (t) # Tag v1.7.0 and push git commits/tags
  release  (r) # Tag v1.7.0 and push built ggem-1.7.0.gem to https://rubygems.org
```

The new version not only lists the cmd names but also any aliases
plus a one-line summary description.  Also the commands are listed
in a custom order (not alphabetical).  The motivation for this
change comes from using bundler's rake gem tasks.  Due to the nature
of rake, you can run `rake -T` and get a similar task listing.  I
found myself relying on this to quickly remind myself of the commands
and to quickly verify the gem source I am about to release to.

To implement this, I created a custom command set class to store the
configured commands.  The `add` method allows specifying any aliases
right alongside the name.  The class keeps track on add order and
lists the commands in this order in its `to_s` representation.

Also included is each commands summary.  The commands have been updated
to break out the summary from the help message and I added a summary
for the generate command.

All of this provides the rich output I have become accustomed to using
Bundler's rake tasks.