(hotfix) generate test helper with Array#sample backfill for 1.8.7
Modern versions of Ruby provide `sample` while the equivalent logic
is called `choice` in 1.8.7.  This allows gem developers to use
this logic in their test suites and have it work in modern Ruby
versions as well as 1.8.7.

This also updates the template to use the latest patch of Assert.
This version is fully compatible with modern Ruby and plus GGem
should encourage using the latest and greatest.