cli: add `force-tag` option to the tag and release commands
By default, the tag and release commands will exit out if there
are uncommitted files.  This adds a `force-tag` option to each
that allows users to override this behavior and proceed with the
command behavior even if there are uncommitted files.

I came across this need while attempting to publish a new version
of the much-plugin gem.  This gem is a dependency of GGem and so
GGem never sees the updated version info b/c the dependency version
info is already loaded.  I'd like to manually override the version
require just to allow releasing the new version but when I override
the gemspec version require, ggem won't let me release b/c there
are uncommitted files.  This allows me to proceed with the release
in this special case.  I want to stress that this special case only
applies to the much-plugin and scmd gems as they are dependencies
of ggem - this is not normal and won't be needed for running ggem
commands on any other gem.