(hotfix) make the push cmd and force tag option cmd
This is necessary b/c the release cmd, which is a force tag option
cmd, calls both the tag and push cmds with the same argv.  Therefore
both tag and push should expect the same argv.  Without this, the
tag part of release works but the push fails saying "unknown option".

I ran into this when trying to use the force option when releasing
a gem.