(hotfix) run sub-cmds with specific argv
This updates the sub-cmd running to be smart about the argv that is
used.  This is needed b/c the argv for the parent cmd may not be
valid for the child cmds.  This is the case for the release cmd
which uses the force-tag option.  It has two child cmds: tag and
push.  The force-tag option is valid for tag but not for push.
This caused the release cmd to always error on push if the force-tag
option was used.

This switches to not blindly running child cmds with the parent's
argv.  The release cmd now intelligently passes on the force-tag
option to the tag cmd if needed.