setup CircleCI test runs
Playing around with this.  Trying to see if I can get it working.

Note: I added gitkeeps for `/tmp` and `/log` (these should have
been in place already and may cause tests to fail if the dirs are
not in place).

Note also: I updated the gemspec to ignore any files starting with
`.` when building the gem.  There's no need to package these files
when pushing gems.  They are only needed for local development and
testing.